+++ modules/home/programs/nixvim/config/snacks.nix
@@ -23,6 +23,7 @@
         enabled = true;
         sources = {
           files = { hidden = true; };
+          explorer = { hidden = true; };
         };
       };
       quickfile = { enabled = true; };