+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Feat
 
+- **ai-tools**: switched from nix-ai-tools to llm-agents flake, moved tools to system packages, added crush, kilocode-cli, mistral-vibe
 - **darwin**: fish and ghostty
 - **nixvim**: ported neovim
 - **nixvim**: pbcopy & colorscheme